+++ src/main/java/com/depromeet/global/security/JwtAuthenticationFilter.java
@@ -3,19 +3,23 @@ package com.depromeet.global.security;
 import static com.depromeet.global.common.constants.SecurityConstants.*;
 
 import com.depromeet.domain.auth.application.JwtTokenService;
-import com.depromeet.domain.auth.dto.AccessToken;
+import com.depromeet.domain.auth.dto.response.AccessToken;
+
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Optional;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+import java.util.Optional;
 
 @Slf4j
 @Component
